swisstransport api

+++ SwissTransport/main/java/src/rest/ConnectionRequest.java
@@ -93,7 +93,7 @@ public class ConnectionRequest extends RestRequest<Connection, ConnectionList> {
 	 *            1 - 6. Specifies the number of connections to return. If several
 	 *            connections depart at the same time they are counted as 1.
 	 */
-	public void limit(byte limit) {
+	public void limit(int limit) {
 		addParam(new Parameter("limit", String.valueOf(limit)));
 	}
 
@@ -103,7 +103,7 @@ public class ConnectionRequest extends RestRequest<Connection, ConnectionList> {
 	 *            0 - 3. Allows pagination of connections. Zero-based, so first page
 	 *            is 0, second is 1, third is 2 and so on.
 	 */
-	public void page(byte page) {
+	public void page(int page) {
 		addParam(new Parameter("page", String.valueOf(page)));
 	}
